Relocating messaging architecture to a different module.
