+++ smartcosmos-framework-messaging/pom.xml
@@ -9,14 +9,18 @@
         <relativePath>../smartcosmos-framework-parent/pom.xml</relativePath>
     </parent>
 
-    <artifactId>smartcosmos-framework</artifactId>
+    <artifactId>smartcosmos-framework-messaging</artifactId>
 
-    <name>SMART COSMOS Framework Annotations</name>
-    <description>Provides Spring Configuration and Annotations making it easier to develop SMART COSMOS Extensions</description>
+    <name>SMART COSMOS Framework Messaging Annotations</name>
+    <description>Provides Spring Configuration and Annotations making it easier to develop SMART COSMOS Extensions that tie into events and asynchronous messaging.</description>
     <dependencies>
         <dependency>
             <groupId>joda-time</groupId>
             <artifactId>joda-time</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>net.smartcosmos</groupId>
+            <artifactId>smartcosmos-framework</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -41,6 +45,12 @@
             <artifactId>spring-cloud-starter-oauth2</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.springframework.integration</groupId>
+            <artifactId>spring-integration-kafka</artifactId>
+            <version>2.0.0.RC1</version>
+            <optional>true</optional>
+        </dependency>
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-integration</artifactId>
         </dependency>
@@ -48,6 +58,14 @@
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
             <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-stream</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-starter-stream-kafka</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
