+++ smartcosmos-framework-messaging/src/main/java/net/smartcosmos/events/EnableSmartCosmosEvents.java
