+++ smartcosmos-framework-messaging/src/main/java/net/smartcosmos/events/SmartCosmosEventConfiguration.java
@@ -1,8 +1,12 @@
 package net.smartcosmos.events;
 
+import org.springframework.context.annotation.Configuration;
+
 /**
  * @author voor
  */
+@Configuration
 public class SmartCosmosEventConfiguration {
+
 
 }
