+++ smartcosmos-framework-messaging/src/main/java/net/smartcosmos/events/SmartCosmosEventSink.java
