+++ smartcosmos-framework-messaging/src/main/java/net/smartcosmos/events/impl/MessagingSmartCosmosEventTemplate.java
@@ -1,14 +1,14 @@
 package net.smartcosmos.events.impl;
 
-import net.smartcosmos.events.SmartCosmosEventProducer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.integration.support.MessageBuilder;
 import org.springframework.stereotype.Component;
 
-import net.smartcosmos.events.ISmartCosmosEventGateway;
+import net.smartcosmos.events.AbstractSmartCosmosEventTemplate;
 import net.smartcosmos.events.SmartCosmosEvent;
 import net.smartcosmos.events.SmartCosmosEventException;
+import net.smartcosmos.events.SmartCosmosEventGateway;
 
 /**
  * @author voor
@@ -16,14 +16,14 @@ import net.smartcosmos.events.SmartCosmosEventException;
 @Component
 public class MessagingSmartCosmosEventTemplate extends AbstractSmartCosmosEventTemplate {
 
-    private final ISmartCosmosEventGateway smartCosmosEventGateway;
+    private final SmartCosmosEventGateway smartCosmosEventGateway;
 
-    @Value("${smart-cosmos.event-header}")
+    @Value("${smartcosmos.event-header}")
     private String smartCosmosEventSource;
 
     @Autowired
     public MessagingSmartCosmosEventTemplate(
-            ISmartCosmosEventGateway smartCosmosEventGateway) {
+            SmartCosmosEventGateway smartCosmosEventGateway) {
         this.smartCosmosEventGateway = smartCosmosEventGateway;
     }
