+++ smartcosmos-framework/src/main/java/net/smartcosmos/events/AbstractSmartCosmosEventTemplate.java
@@ -1,15 +1,12 @@
-package net.smartcosmos.events.impl;
+package net.smartcosmos.events;
 
-import net.smartcosmos.events.ISmartCosmosEventTemplate;
-import net.smartcosmos.events.SmartCosmosEvent;
-import net.smartcosmos.events.SmartCosmosEventException;
 import net.smartcosmos.security.user.SmartCosmosUser;
 
 /**
  * @author voor
  */
 public abstract class AbstractSmartCosmosEventTemplate
-        implements ISmartCosmosEventTemplate {
+        implements SmartCosmosEventTemplate {
 
     public abstract void convertAndSend(SmartCosmosEvent<Object> message)
             throws SmartCosmosEventException;
